Move API save tests to controller tests
[MAILPOET-2900]
